+++ packages/contracts/contracts/HintHelpers.sol
@@ -61,21 +61,21 @@ contract HintHelpers is LiquityBase, HintHelpersStorage, CheckContract {
      */
 
     function getRedemptionHints(
-        uint _ZUSDamount, 
-        uint _price,
-        uint _maxIterations
+        uint256 _ZUSDamount, 
+        uint256 _price,
+        uint256 _maxIterations
     )
         external
         view
         returns (
             address firstRedemptionHint,
-            uint partialRedemptionHintNICR,
-            uint truncatedZUSDamount
+            uint256 partialRedemptionHintNICR,
+            uint256 truncatedZUSDamount
         )
     {
         ISortedTroves sortedTrovesCached = sortedTroves;
 
-        uint remainingZUSD = _ZUSDamount;
+        uint256 remainingZUSD = _ZUSDamount;
         address currentTroveuser = sortedTrovesCached.getLast();
 
         while (currentTroveuser != address(0) && troveManager.getCurrentICR(currentTroveuser, _price) < liquityBaseParams.MCR()) {
@@ -89,20 +89,20 @@ contract HintHelpers is LiquityBase, HintHelpersStorage, CheckContract {
         }
 
         while (currentTroveuser != address(0) && remainingZUSD > 0 && _maxIterations-- > 0) {
-            uint netZUSDDebt = _getNetDebt(troveManager.getTroveDebt(currentTroveuser))
+            uint256 netZUSDDebt = _getNetDebt(troveManager.getTroveDebt(currentTroveuser))
                 .add(troveManager.getPendingZUSDDebtReward(currentTroveuser));
 
             if (netZUSDDebt > remainingZUSD) {
                 if (netZUSDDebt > MIN_NET_DEBT) {
-                    uint maxRedeemableZUSD = LiquityMath._min(remainingZUSD, netZUSDDebt.sub(MIN_NET_DEBT));
+                    uint256 maxRedeemableZUSD = LiquityMath._min(remainingZUSD, netZUSDDebt.sub(MIN_NET_DEBT));
 
-                    uint RBTC = troveManager.getTroveColl(currentTroveuser)
+                    uint256 RBTC = troveManager.getTroveColl(currentTroveuser)
                         .add(troveManager.getPendingRBTCReward(currentTroveuser));
 
-                    uint newColl = RBTC.sub(maxRedeemableZUSD.mul(DECIMAL_PRECISION).div(_price));
-                    uint newDebt = netZUSDDebt.sub(maxRedeemableZUSD);
+                    uint256 newColl = RBTC.sub(maxRedeemableZUSD.mul(DECIMAL_PRECISION).div(_price));
+                    uint256 newDebt = netZUSDDebt.sub(maxRedeemableZUSD);
 
-                    uint compositeDebt = _getCompositeDebt(newDebt);
+                    uint256 compositeDebt = _getCompositeDebt(newDebt);
                     partialRedemptionHintNICR = LiquityMath._computeNominalCR(newColl, compositeDebt);
 
                     remainingZUSD = remainingZUSD.sub(maxRedeemableZUSD);
@@ -127,12 +127,12 @@ contract HintHelpers is LiquityBase, HintHelpersStorage, CheckContract {
     Submitting numTrials = k * sqrt(length), with k = 15 makes it very, very likely that the ouput address will 
     be <= sqrt(length) positions away from the correct insert position.
     */
-    function getApproxHint(uint _CR, uint _numTrials, uint _inputRandomSeed)
+    function getApproxHint(uint256 _CR, uint256 _numTrials, uint256 _inputRandomSeed)
         external
         view
-        returns (address hintAddress, uint diff, uint latestRandomSeed)
+        returns (address hintAddress, uint256 diff, uint256 latestRandomSeed)
     {
-        uint arrayLength = troveManager.getTroveOwnersCount();
+        uint256 arrayLength = troveManager.getTroveOwnersCount();
 
         if (arrayLength == 0) {
             return (address(0), 0, _inputRandomSeed);
@@ -142,17 +142,17 @@ contract HintHelpers is LiquityBase, HintHelpersStorage, CheckContract {
         diff = LiquityMath._getAbsoluteDifference(_CR, troveManager.getNominalICR(hintAddress));
         latestRandomSeed = _inputRandomSeed;
 
-        uint i = 1;
+        uint256 i = 1;
 
         while (i < _numTrials) {
             latestRandomSeed = uint(keccak256(abi.encodePacked(latestRandomSeed)));
 
-            uint arrayIndex = latestRandomSeed % arrayLength;
+            uint256 arrayIndex = latestRandomSeed % arrayLength;
             address currentAddress = troveManager.getTroveFromTroveOwnersArray(arrayIndex);
-            uint currentNICR = troveManager.getNominalICR(currentAddress);
+            uint256 currentNICR = troveManager.getNominalICR(currentAddress);
 
             // check if abs(current - CR) > abs(closest - CR), and update closest if current is closer
-            uint currentDiff = LiquityMath._getAbsoluteDifference(currentNICR, _CR);
+            uint256 currentDiff = LiquityMath._getAbsoluteDifference(currentNICR, _CR);
 
             if (currentDiff < diff) {
                 diff = currentDiff;
@@ -162,11 +162,11 @@ contract HintHelpers is LiquityBase, HintHelpersStorage, CheckContract {
         }
     }
 
-    function computeNominalCR(uint _coll, uint _debt) external pure returns (uint) {
+    function computeNominalCR(uint256 _coll, uint256 _debt) external pure returns (uint) {
         return LiquityMath._computeNominalCR(_coll, _debt);
     }
 
-    function computeCR(uint _coll, uint _debt, uint _price) external pure returns (uint) {
+    function computeCR(uint256 _coll, uint256 _debt, uint256 _price) external pure returns (uint) {
         return LiquityMath._computeCR(_coll, _debt, _price);
     }
 }
