+++ packages/contracts/contracts/Interfaces/ITroveManager.sol
@@ -29,17 +29,17 @@ interface ITroveManager is ILiquityBase {
     event ZEROTokenAddressChanged(address _zeroTokenAddress);
     event ZEROStakingAddressChanged(address _zeroStakingAddress);
 
-    event Liquidation(uint _liquidatedDebt, uint _liquidatedColl, uint _collGasCompensation, uint _ZUSDGasCompensation);
-    event Redemption(uint _attemptedZUSDAmount, uint _actualZUSDAmount, uint _RBTCSent, uint _RBTCFee);
-    event TroveUpdated(address indexed _borrower, uint _debt, uint _coll, uint stake, uint8 operation);
-    event TroveLiquidated(address indexed _borrower, uint _debt, uint _coll, uint8 operation);
-    event BaseRateUpdated(uint _baseRate);
-    event LastFeeOpTimeUpdated(uint _lastFeeOpTime);
-    event TotalStakesUpdated(uint _newTotalStakes);
-    event SystemSnapshotsUpdated(uint _totalStakesSnapshot, uint _totalCollateralSnapshot);
-    event LTermsUpdated(uint _L_RBTC, uint _L_ZUSDDebt);
-    event TroveSnapshotsUpdated(uint _L_RBTC, uint _L_ZUSDDebt);
-    event TroveIndexUpdated(address _borrower, uint _newIndex);
+    event Liquidation(uint256 _liquidatedDebt, uint256 _liquidatedColl, uint256 _collGasCompensation, uint256 _ZUSDGasCompensation);
+    event Redemption(uint256 _attemptedZUSDAmount, uint256 _actualZUSDAmount, uint256 _RBTCSent, uint256 _RBTCFee);
+    event TroveUpdated(address indexed _borrower, uint256 _debt, uint256 _coll, uint256 stake, uint8 operation);
+    event TroveLiquidated(address indexed _borrower, uint256 _debt, uint256 _coll, uint8 operation);
+    event BaseRateUpdated(uint256 _baseRate);
+    event LastFeeOpTimeUpdated(uint256 _lastFeeOpTime);
+    event TotalStakesUpdated(uint256 _newTotalStakes);
+    event SystemSnapshotsUpdated(uint256 _totalStakesSnapshot, uint256 _totalCollateralSnapshot);
+    event LTermsUpdated(uint256 _L_RBTC, uint256 _L_ZUSDDebt);
+    event TroveSnapshotsUpdated(uint256 _L_RBTC, uint256 _L_ZUSDDebt);
+    event TroveIndexUpdated(address _borrower, uint256 _newIndex);
 
     // --- Functions ---
     /**
@@ -69,7 +69,7 @@ interface ITroveManager is ILiquityBase {
 
     /// @param _index Trove owner index
     /// @return Trove from TroveOwners array in given index
-    function getTroveFromTroveOwnersArray(uint _index) external view returns (address);
+    function getTroveFromTroveOwnersArray(uint256 _index) external view returns (address);
 
     /// @param _borrower borrower address
     /// @return the nominal collateral ratio (ICR) of a given Trove, without the price. Takes a trove's pending coll and debt rewards from redistributions into account.
@@ -79,7 +79,7 @@ interface ITroveManager is ILiquityBase {
     /// @param _borrower borrower address
     /// @param _price RBTC price
     /// @return the current collateral ratio (ICR) of a given Trove. Takes a trove's pending coll and debt rewards from redistributions into account.
-    function getCurrentICR(address _borrower, uint _price) external view returns (uint);
+    function getCurrentICR(address _borrower, uint256 _price) external view returns (uint);
 
     /// @notice Closes the trove if its ICR is lower than the minimum collateral ratio.
     /// @param _borrower borrower address
@@ -90,7 +90,7 @@ interface ITroveManager is ILiquityBase {
      * starting from the one with the lowest collateral ratio in the system, and moving upwards
      * @param _n max number of under-collateralized Troves to liquidate
      */
-    function liquidateTroves(uint _n) external;
+    function liquidateTroves(uint256 _n) external;
 
     /**
      * @notice Attempt to liquidate a custom list of troves provided by the caller.
@@ -126,13 +126,13 @@ interface ITroveManager is ILiquityBase {
      * @param _maxFee max fee percentage to accept
      */
     function redeemCollateral(
-        uint _ZUSDAmount,
+        uint256 _ZUSDAmount,
         address _firstRedemptionHint,
         address _upperPartialRedemptionHint,
         address _lowerPartialRedemptionHint,
-        uint _partialRedemptionHintNICR,
-        uint _maxIterations,
-        uint _maxFee
+        uint256 _partialRedemptionHintNICR,
+        uint256 _maxIterations,
+        uint256 _maxFee
     ) external; 
 
     /// @notice Update borrower's stake based on their latest collateral value
@@ -146,7 +146,7 @@ interface ITroveManager is ILiquityBase {
     /// @notice Push the owner's address to the Trove owners list, and record the corresponding array index on the Trove struct
     /// @param _borrower borrower address
     /// @return index where Trove was inserted
-    function addTroveOwnerToArray(address _borrower) external returns (uint index);
+    function addTroveOwnerToArray(address _borrower) external returns (uint256 index);
 
     /// @notice Add the borrowers's coll and debt rewards earned from redistributions, to their Trove
     /// @param _borrower borrower address
@@ -173,10 +173,10 @@ interface ITroveManager is ILiquityBase {
     /// @notice returns the Troves entire debt and coll, including pending rewards from redistributions.
     /// @param _borrower borrower address
     function getEntireDebtAndColl(address _borrower) external view returns (
-        uint debt, 
-        uint coll, 
-        uint pendingZUSDDebtReward, 
-        uint pendingRBTCReward
+        uint256 debt, 
+        uint256 coll, 
+        uint256 pendingZUSDDebtReward, 
+        uint256 pendingRBTCReward
     );
 
     /// @notice Close given trove. Called by BorrowerOperations.
@@ -195,7 +195,7 @@ interface ITroveManager is ILiquityBase {
 
     /// @notice The redemption fee is taken as a cut of the total RBTC drawn from the system in a redemption. It is based on the current redemption rate.
     /// @param _RBTCDrawn RBTC drawn
-    function getRedemptionFeeWithDecay(uint _RBTCDrawn) external view returns (uint);
+    function getRedemptionFeeWithDecay(uint256 _RBTCDrawn) external view returns (uint);
 
     /// @return borrowing rate
     function getBorrowingRate() external view returns (uint);
@@ -205,11 +205,11 @@ interface ITroveManager is ILiquityBase {
 
     /// @param ZUSDDebt ZUSD debt amount to calculate fee
     /// @return borrowing fee using borrowing rate
-    function getBorrowingFee(uint ZUSDDebt) external view returns (uint);
+    function getBorrowingFee(uint256 ZUSDDebt) external view returns (uint);
 
     /// @param _ZUSDDebt ZUSD debt amount to calculate fee
     /// @return borrowing fee using borrowing rate with decay
-    function getBorrowingFeeWithDecay(uint _ZUSDDebt) external view returns (uint);
+    function getBorrowingFeeWithDecay(uint256 _ZUSDDebt) external view returns (uint);
 
     /// @notice Updates the baseRate state variable based on time elapsed since the last redemption or ZUSD borrowing operation.
     function decayBaseRateFromBorrowing() external;
@@ -232,39 +232,39 @@ interface ITroveManager is ILiquityBase {
 
     /// @param _borrower borrower address
     /// @param num status to set
-    function setTroveStatus(address _borrower, uint num) external;
+    function setTroveStatus(address _borrower, uint256 num) external;
 
     /// @param _borrower borrower address
     /// @param _collIncrease amount of collateral to increase
     /// @return new trove collateral
-    function increaseTroveColl(address _borrower, uint _collIncrease) external returns (uint);
+    function increaseTroveColl(address _borrower, uint256 _collIncrease) external returns (uint);
 
     /// @param _borrower borrower address
     /// @param _collDecrease amount of collateral to decrease
     /// @return new trove collateral
-    function decreaseTroveColl(address _borrower, uint _collDecrease) external returns (uint); 
+    function decreaseTroveColl(address _borrower, uint256 _collDecrease) external returns (uint); 
 
     /// @param _borrower borrower address
     /// @param _debtIncrease amount of debt to increase
     /// @return new trove debt
-    function increaseTroveDebt(address _borrower, uint _debtIncrease) external returns (uint); 
+    function increaseTroveDebt(address _borrower, uint256 _debtIncrease) external returns (uint); 
 
     /// @param _borrower borrower address
     /// @param _debtDecrease amount of debt to decrease
     /// @return new trove debt
-    function decreaseTroveDebt(address _borrower, uint _debtDecrease) external returns (uint); 
+    function decreaseTroveDebt(address _borrower, uint256 _debtDecrease) external returns (uint); 
 
     /**
      * @param _price RBTC price
      * @return the total collateralization ratio (TCR) of the system. 
      * The TCR is based on the the entire system debt and collateral (including pending rewards).
      */
-    function getTCR(uint _price) external view returns (uint);
+    function getTCR(uint256 _price) external view returns (uint);
 
     function MCR() external view returns (uint);
 
     function CCR() external view returns (uint);
 
     /// @notice reveals whether or not the system is in Recovery Mode (i.e. whether the Total Collateralization Ratio (TCR) is below the Critical Collateralization Ratio (CCR)).
-    function checkRecoveryMode(uint _price) external view returns (bool);
+    function checkRecoveryMode(uint256 _price) external view returns (bool);
 }
