+++ packages/contracts/contracts/Interfaces/IZUSDToken.sol
@@ -13,7 +13,7 @@ interface IZUSDToken is IERC20, IERC2612 {
     event StabilityPoolAddressChanged(address _newStabilityPoolAddress);
     event BorrowerOperationsAddressChanged(address _newBorrowerOperationsAddress);
 
-    event ZUSDTokenBalanceUpdated(address _user, uint _amount);
+    event ZUSDTokenBalanceUpdated(address _user, uint256 _amount);
 
     // --- Functions ---
